Added ownership and admin permission.

+++ controllers/order.go
@@ -3,7 +3,6 @@ package controllers
 import (
 	"fmt"
 	"net/http"
-
 	"github.com/alfredamos/middlewares"
 	"github.com/alfredamos/models"
 	"github.com/gin-gonic/gin"
@@ -42,6 +41,14 @@ func DeleteOrderById(context *gin.Context){
 
 	//----> Get order id from params.
 	id := context.Param("id")
+
+	//----> Check for ownership or admin privilege.
+	isAdmin, isOwnership := checkForOwnershipAndAdmin(id, context)
+
+	if !isAdmin && !isOwnership {
+		context.JSON(http.StatusUnauthorized, gin.H{"status": "failed", "message": "You are not authorized to delete this order!"})
+		return
+	}
 	
 	//----> Delete order with this id.
 	err := order.DeleteOrderById(id)
@@ -64,7 +71,7 @@ func DeleteOrderByUserId(context *gin.Context){
 	userId := context.Param("userId")
 
 	//----> Check for ownership permission or admin privilege.
-	err := middlewares.OwnerAuthorize(userId, context)
+	err := middlewares.CheckForSameUserAndAdmin(userId, context)
 
 	//----> Check for ownership.
 	if err != nil {
@@ -127,7 +134,7 @@ func GetAllOrderByUserId(context *gin.Context){
 	userId := context.Param("userId")
 
 	//----> Check for ownership permission or admin privilege.
-	err := middlewares.OwnerAuthorize(userId, context)
+	err := middlewares.CheckForSameUserAndAdmin(userId, context)
 	
 	//----> Check for error.
 	if err != nil {
@@ -152,20 +159,19 @@ func GetOrderById(context *gin.Context){
 	//----> declare the order variable.
 	order := models.Order{}
 
-	//----> Check for ownership permission or admin privilege.
-	err := middlewares.OwnerAuthorize(order.UserID, context)
-
-	//----> Check for ownership.
-	if err != nil {
-		context.JSON(http.StatusForbidden, gin.H{"status": "failed!", "message": fmt.Sprintf("%v", err)})
-		return
-	}
-
 	//----> The id from params.
 	id := context.Param("id")
 
+	//----> Check for ownership or admin privilege.
+	isAdmin, isOwnership := checkForOwnershipAndAdmin(id, context)
+
+	if !isAdmin && !isOwnership {
+		context.JSON(http.StatusUnauthorized, gin.H{"status": "failed", "message": "You are not authorized to delete this order!"})
+		return
+	}
+
 	//----> Get order by order-id.
-	order, err = order.GetOrderById(id)
+	order, err := order.GetOrderById(id)
 
 	//----> Check for error.
 	if err != nil {
